% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDM.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{Create distance matrix}
\usage{
create_distance_matrix(df)
}
\arguments{
\item{df}{data frame including cases (point1,point2,distance)}
}
\value{
distance matrix
}
\description{
This function can create distance matrix from data frame with proper
structure.
}
\details{
Distances between points should be defined in form of distance matrix (see
vignette for package for details). create_distance_matrix() function uses as
an input  data frame for which every row includes case: ("pointX","pointY",
distance_between_them). As an output proper distance matrix is generated.

Format of input data frame:

Format 1
\itemize{
\item first column - numeric
\item second column - numeric
\item third column - numeric
}
For such case we assume that n points are enumerated by integers 1 .. n and
first two columns include those indexes.

Format 2
\itemize{
\item first column - character
\item second column - character
\item third column - numeric
}
For such case we are using points names instead of indexes. All used names
are sorted and than indexes (used in output distance matrix) will be
assigned to following points.

When input data frame is not complete (not all possible distances are
described) missing cases will be marked in distance matrix as NA.

create_distance_matrix secures first and second metrics rule (see vignette
for package) so even if this info is not included in input df output dm
will include 0 in diagonal.
When there are many (different) distances for the same points in data frame,
output distance matrix will include only last (for both (i,j) and (j,i)
cases - so second metrics rule is fulfilled).
}
