% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDM.R
\name{create_dm_from_df_format_1}
\alias{create_dm_from_df_format_1}
\title{Function takes as an input data frame prepared in format 1
(column types - numeric, numeric, numeric) and returns distance matrix}
\usage{
create_dm_from_df_format_1(df)
}
\description{
Function takes as an input data frame prepared in format 1
(column types - numeric, numeric, numeric) and returns distance matrix
}
