% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDM.R
\name{create_dm_from_df_format_2}
\alias{create_dm_from_df_format_2}
\title{Function takes as an input data frame prepared in format 2
(column types - character, character, numeric) and returns distance matrix}
\usage{
create_dm_from_df_format_2(df)
}
\description{
Function takes as an input data frame prepared in format 2
(column types - character, character, numeric) and returns distance matrix
}
