% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{debug_distance_matrix}
\alias{debug_distance_matrix}
\title{Debug distance matrix}
\usage{
debug_distance_matrix(m)
}
\arguments{
\item{m}{Distance Matrix which is to be validate}
}
\description{
The debug_distance_matrix() function is used for troubleshooting.
}
\details{
If distance matrix is to be used by to find points, it has to be well
defined. This function checks if matrix is square, if doesn't include NA or
negative values. If those conditions are not fulfilled proper information is
printed.

Additionally metric rules are checked:

\enumerate{
 \item d(x,y) = 0 <=> x = y
 \item d(x,y) = d(y,x)
 \item d(x,y) <= d(x,z) + d(y,z)
}
If some rule is not fulfilled indexes of points/distance which cause
problems are printed
}
