% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{debug_matrix_for_metric_rule1}
\alias{debug_matrix_for_metric_rule1}
\title{Function debug matrix content according to first metric rule
 d(x,y) = 0 <=> x = y}
\usage{
debug_matrix_for_metric_rule1(m)
}
\description{
Printout returned includes information about point which cause problems.
If first metric rule is fulfilled, empty string is returned.
}
