% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{debug_matrix_for_metric_rule2}
\alias{debug_matrix_for_metric_rule2}
\title{Function debug matrix content according to second metric rule
 d(x,y) = d(y,x)}
\usage{
debug_matrix_for_metric_rule2(m)
}
\description{
Printout returned includes information about pair of point which
cause problems.
If second metric rule is fulfilled, empty string is returned.
}
