% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{debug_matrix_for_metric_rule3}
\alias{debug_matrix_for_metric_rule3}
\title{Function debug matrix content according to third metric rule
d(x,z) <= d(x,y) + d(x,z)}
\usage{
debug_matrix_for_metric_rule3(m)
}
\description{
Printout returned includes information about triples of point which
cause problems.
If third metric rule is fulfilled, empty string is returned.
}
\details{
Warning: algorithm works properly when rule2 is fulfilled
}
