% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairDM.R
\name{fix.na}
\alias{fix.na}
\title{Function removes all NA from distance matrix}
\usage{
fix.na(m)
}
\arguments{
\item{m}{Distance Matrix}
}
\value{
distance matrix without NA
}
\description{
NA on diagonal replaced by 0
NA for (i,j) replaced by (j,i) if exist
otherwise replaced by meand from existing number (or by 1 if they don't exist)
}
