% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{is_distance_matrix_ok}
\alias{is_distance_matrix_ok}
\title{Check if matrix fulfills distance matrix conditions}
\usage{
is_distance_matrix_ok(m)
}
\arguments{
\item{m}{Distance Matrix which is to be validate}
}
\value{
If all conditions are fulfilled TRUE is returned, else FALSE.
}
\description{
The `is_distance_matrix_ok()` function is used to validate matrix.
}
\details{
There are distance matrix conditions which should be fulfilled:

3 distance rules
\enumerate{
 \item d(x,y) = 0 <=> x = y
 \item d(x,y) = d(y,x)
 \item d(x,y) <= d(x,z) + d(y,z)
}
Additionally matrix should have the same number of rows and columns,
and negative or NA values should not be included.
}
