% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{is_metric_rule1_fulfilled}
\alias{is_metric_rule1_fulfilled}
\title{Function checks if first metric rule is fulfilled
 d(x,y) = 0 <=> x = y}
\usage{
is_metric_rule1_fulfilled(m)
}
\description{
Function checks if first metric rule is fulfilled
 d(x,y) = 0 <=> x = y
}
