% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{is_metric_rule2_fulfilled}
\alias{is_metric_rule2_fulfilled}
\title{Function checks if second metric rule is fulfilled
d(x,y) = d(y,x)}
\usage{
is_metric_rule2_fulfilled(m)
}
\description{
Function checks if second metric rule is fulfilled
d(x,y) = d(y,x)
}
