% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{is_metric_rule3_fulfilled}
\alias{is_metric_rule3_fulfilled}
\title{Function checks if third metric rule is fulfilled
d(x,z) <= d(x,y) + d(x,z)
Warning: algorithm works properly when rule2 is fulfilled}
\usage{
is_metric_rule3_fulfilled(m)
}
\description{
Function checks if third metric rule is fulfilled
d(x,z) <= d(x,y) + d(x,z)
Warning: algorithm works properly when rule2 is fulfilled
}
