% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDM.R
\name{is_square_matrix}
\alias{is_square_matrix}
\title{Function returns TRUE for square matrix, else FALSE is returned}
\usage{
is_square_matrix(m)
}
\description{
Function returns TRUE for square matrix, else FALSE is returned
}
