% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairDM.R
\name{make_square_matrix}
\alias{make_square_matrix}
\title{Function adds columns or rows to make matrix square}
\usage{
make_square_matrix(m)
}
\description{
Function adds columns or rows to make matrix square
}
