% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairDM.R
\name{repair_distance_matrix}
\alias{repair_distance_matrix}
\title{Repair distance matrix}
\usage{
repair_distance_matrix(m, max.iter = 5)
}
\arguments{
\item{m}{original distance matrix}

\item{max.iter}{= 5, repairing third metric rule can be a little bit tricky
when there are many triples combination which don't fulfill it. In such
situation many iterations are needed. If number of needed iteration exceeds
max.iter, simple matrix will be returned when all possible distances between
points will be equal to the mean distance in original matrix.}
}
\value{
repaired distance matrix (or original if repairing is not needed)
}
\description{
This function can repair distance matrix to satisfy all conditions needed
for  proper input for algorithms which are looking for points positions.
}
\details{
\itemize{
\item If matrix is not square, it will be extended to fulfill this condition
\item All NA will be replaced by real values
\item All negative values will be changed by abs function
\item First metric rule will be forced (d(x,y) = 0 <=> x = y)
\item Second metric rule will be forced (d(x,y) = d(y,x))
\item Third metric rule will be forced (d(x,z) <= d(x,y) + d(y,z))
}
All changes are performed in intelligent way to change original matrix as
less as possible
}
