% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairDM.R
\name{repair_metric_rule1}
\alias{repair_metric_rule1}
\title{Function repairs dimension matrix to fulfill
first metric rule d(x,y) = 0 <=> x = y
Remark: there is an assumption that NA are not included (fix.na() used first)}
\usage{
repair_metric_rule1(m)
}
\description{
Function repairs dimension matrix to fulfill
first metric rule d(x,y) = 0 <=> x = y
Remark: there is an assumption that NA are not included (fix.na() used first)
}
