% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairDM.R
\name{repair_metric_rule2}
\alias{repair_metric_rule2}
\title{Function repairs dimension matrix to fulfill
second metric rule d(x,y) = d(y,x)}
\usage{
repair_metric_rule2(m)
}
\description{
Upper triangular part overwrites lower triangular part of matrix
}
