% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repairDM.R
\name{repair_metric_rule3}
\alias{repair_metric_rule3}
\title{Function repairs dimension matrix to fulfill
third metric rule d(x,z) <= d(x,y) + d(y,z)
Warning: algorithm works properly when rule2 is fulfilled}
\usage{
repair_metric_rule3(m, max.iter = 5)
}
\description{
@param m Distance Matrix
 @param max.iter = 5, repairing third metric rule can be a little bit tricky
 when there are many triples combination which don't fulfill it. In such
 situation many iteration is needed. If number of needed iteration exceeds
 max.iter, simple matrix will be returned when all possible distances between
 points will be equal to the mean distance in original matrix.
}
