% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDM.R
\name{validate_df}
\alias{validate_df}
\title{Function validates data frame with distance cases (point1,point2,distance)
\itemize{
\item If values included in all three columns are numeric 1 is returned
\item If values included in first and second column are character and third
includes numeric values 2 is returned
}
Otherwise, this is not a valid data frame, error is generated}
\usage{
validate_df(df)
}
\description{
Function validates data frame with distance cases (point1,point2,distance)
\itemize{
\item If values included in all three columns are numeric 1 is returned
\item If values included in first and second column are character and third
includes numeric values 2 is returned
}
Otherwise, this is not a valid data frame, error is generated
}
